Move "Attach file" to Insert menu
Move "Attach file" to the insert menu and fix regression
by inserting link on insert.

Fixes #744
