+++ tests/uiactions.py
@@ -608,70 +608,6 @@ class TestUIActions(tests.TestCase):
 
 		self.assertEqual(self.navigation.opened, Path('NewPage'))
 
-	def testAttachFile(self):
-		folder = self.setUpFolder(mock=tests.MOCK_ALWAYS_REAL)
-		file = folder.file('Attachment.abc')
-		file.write('Test ABC\n')
-
-		def attach_file(dialog):
-			dialog.set_file(file)
-			dialog.assert_response_ok()
-
-		with tests.DialogContext(attach_file):
-			self.uiactions.attach_file()
-
-		attach_folder = self.notebook.get_attachments_dir(self.page)
-		attach_file = attach_folder.file('Attachment.abc')
-		self.assertTrue(attach_file.exists())
-		self.assertEqual(attach_file.read(), file.read())
-
-	def testAttachFileResolveExistingFile(self):
-		folder = self.setUpFolder(mock=tests.MOCK_ALWAYS_REAL)
-		file = folder.file('Attachment.abc')
-		file.write('Test ABC\n')
-
-		attach_folder = self.notebook.get_attachments_dir(self.page)
-		conflict_file = attach_folder.file('Attachment.abc')
-		conflict_file.write('Conflict\n')
-
-		def attach_file(dialog):
-			dialog.set_file(file)
-			dialog.assert_response_ok()
-
-		def resolve_conflict(dialog):
-			dialog.set_input(name='NewName.abc')
-			dialog.assert_response_ok()
-
-		with tests.DialogContext(attach_file, resolve_conflict):
-			self.uiactions.attach_file()
-
-		attach_file = attach_folder.file('NewName.abc')
-		self.assertTrue(attach_file.exists())
-		self.assertEqual(attach_file.read(), file.read())
-
-		self.assertEqual(conflict_file.read(), 'Conflict\n')
-
-	def testAttachFileOverwriteExistingFile(self):
-		folder = self.setUpFolder(mock=tests.MOCK_ALWAYS_REAL)
-		file = folder.file('Attachment.abc')
-		file.write('Test ABC\n')
-
-		attach_folder = self.notebook.get_attachments_dir(self.page)
-		conflict_file = attach_folder.file('Attachment.abc')
-		conflict_file.write('Conflict\n')
-
-		def attach_file(dialog):
-			dialog.set_file(file)
-			dialog.assert_response_ok()
-
-		def resolve_conflict(dialog):
-			dialog.do_response_overwrite()
-
-		with tests.DialogContext(attach_file, resolve_conflict):
-			self.uiactions.attach_file()
-
-		self.assertEqual(conflict_file.read(), file.read())
-
 	def testShowServerDialog(self):
 		from zim.main import ZIM_APPLICATION
 		ZIM_APPLICATION._running = True # HACK
