+++ zim/gui/uiactions.py
@@ -308,14 +308,6 @@ class UIActions(object):
 		dialog = RecentChangesDialog.unique(self, self.widget, self.notebook, self.navigation)
 		dialog.present()
 
-	@action(_('Attach _File'), verb_icon='zim-attachment', menuhints='notebook:edit') # T: Menu item
-	def attach_file(self, path=None):
-		'''Menu action to show the L{AttachFileDialog}
-		@param path: a L{Path} object, or C{None} for the current
-		selected page
-		'''
-		AttachFileDialog(self.widget, self.notebook, path or self.page).run()
-
 	@action(_('Open Attachments _Folder')) # T: Menu item
 	def open_attachments_folder(self):
 		'''Menu action to open the attachment folder for the current page'''
@@ -714,130 +706,6 @@ class DeletePageDialog(Dialog):
 		dialog = ProgressDialog(self, op)
 		dialog.run()
 
-		return True
-
-
-class AttachFileDialog(FileDialog):
-
-	def __init__(self, widget, notebook, path):
-		assert path, 'Need a page here'
-		FileDialog.__init__(self, widget, _('Attach File'), multiple=True) # T: Dialog title
-		self.notebook = notebook
-		self.path = path
-
-		self.add_shortcut(notebook, path)
-		self.load_last_folder()
-
-		dir = notebook.get_attachments_dir(path)
-		if dir is None:
-			ErrorDialog(_('Page "%s" does not have a folder for attachments') % self.path)
-				# T: Error dialog - %s is the full page name
-			raise Exception('Page "%s" does not have a folder for attachments' % self.path)
-
-	def do_response_ok(self):
-		files = self.get_files()
-		if not files:
-			return False
-
-		self.save_last_folder()
-
-		for i, file in enumerate(files):
-			file = attach_file(self, self.notebook, self.path, file)
-			if file is None:
-				return False # overwrite dialog was canceled
-
-		return True
-
-
-def attach_file(widget, notebook, path, file, force_overwrite=False):
-	folder = notebook.get_attachments_dir(path)
-	if folder is None:
-		raise Error('%s does not have an attachments dir' % path)
-
-	dest = folder.file(file.basename)
-	if dest.exists() and not force_overwrite:
-		dialog = PromptExistingFileDialog(widget, dest)
-		dest = dialog.run()
-		if dest is None:
-			return None	# dialog was cancelled
-		elif dest.exists():
-			dest.remove()
-
-	file.copyto(dest)
-	return dest
-
-
-class PromptExistingFileDialog(Dialog):
-	'''Dialog that is used e.g. when a file should be attached to zim,
-	but a file with the same name already exists in the attachment
-	directory. This Dialog allows to suggest a new name or overwrite
-	the existing one.
-
-	For this dialog C{run()} will return either the original file
-	(for overwrite), a new file, or None when the dialog was canceled.
-	'''
-
-	def __init__(self, widget, file):
-		Dialog.__init__(self, widget, _('File Exists'), buttons=None) # T: Dialog title
-		self.add_help_text( _('''\
-A file with the name <b>"%s"</b> already exists.
-You can use another name or overwrite the existing file.''' % file.basename),
-		) # T: Dialog text in 'new filename' dialog
-		self.folder = file.parent()
-		self.old_file = file
-
-		suggested_filename = self.folder.new_file(file.basename).basename
-		self.add_form((
-				('name', 'string', _('Filename')), # T: Input label
-			), {
-				'name': suggested_filename
-			}
-		)
-		self.form.widgets['name'].set_check_func(self._check_valid)
-
-		# all buttons are defined in this class, to get the ordering right
-		# [show folder]      [overwrite] [cancel] [ok]
-		button = Gtk.Button.new_with_mnemonic(_('_Browse')) # T: Button label
-		button.connect('clicked', self.do_show_folder)
-		self.action_area.add(button)
-		self.action_area.set_child_secondary(button, True)
-
-		button = Gtk.Button.new_with_mnemonic(_('Overwrite')) # T: Button label
-		button.connect('clicked', self.do_response_overwrite)
-		self.add_action_widget(button, Gtk.ResponseType.NONE)
-
-		self.add_button(_('_Cancel'), Gtk.ResponseType.CANCEL) # T: Button label
-		self.add_button(_('_OK'), Gtk.ResponseType.OK) # T: Button label
-		self._no_ok_action = False
-
-		self.form.widgets['name'].connect('focus-in-event', self._on_focus)
-
-	def _on_focus(self, widget, event):
-		# filename length without suffix
-		length = len(os.path.splitext(widget.get_text())[0])
-		widget.select_region(0, length)
-
-	def _check_valid(self, filename):
-		# Only valid when same dir and does not yet exist
-		file = self.folder.file(filename)
-		return file.ischild(self.folder) and not file.exists()
-
-	def do_show_folder(self, *a):
-		open_folder(self, self.folder)
-
-	def do_response_overwrite(self, *a):
-		logger.info('Overwriting %s', self.old_file.path)
-		self.result = self.old_file
-
-	def do_response_ok(self):
-		if not self.form.widgets['name'].get_input_valid():
-			return False
-
-		newfile = self.folder.file(self.form['name'])
-		logger.info('Selected %s', newfile.path)
-		assert newfile.ischild(self.folder) # just to be real sure
-		assert not newfile.exists() # just to be real sure
-		self.result = newfile
 		return True
